pyproject.toml
[project]
name = "learn-dspy"
version = "0.1.0"
description = "Learning DSPy framework with practical examples"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy-ai>=3.0.3",
]
